fix: sửa lỗi mọi user đều có thể truy cập DB

+++ actions/deleteChuong.php
@@ -1,9 +1,13 @@
 <?php
 session_start();
+if (!isset($_SESSION['Role']) || $_SESSION['Role'] !== 'admin') {
+        // không phải admin từ chối truy cập
+        header("Location: /403.php"); 
+        exit();
+        }
 include 'connect.php';
 
 $idChuong = (int)$_GET['Id_chuong'];
-
 // Chuẩn bị câu lệnh SQL
 $sql = "DELETE FROM chuong WHERE Id_chuong = ?";
 $stmt = $conn->prepare($sql);
