+++ actions/deleteTruyen.php
@@ -1,9 +1,13 @@
 <?php
 session_start();
+if (!isset($_SESSION['Role']) || $_SESSION['Role'] !== 'admin') {
+        // không phải admin từ chối truy cập
+        header("Location: /403.php"); 
+        exit();
+        }
+
 include 'connect.php';
-
 $id = $_GET['id'];
-
 // Chuẩn bị câu lệnh SQL
 $sql = "DELETE FROM truyen WHERE Id_truyen = ?";
 $stmt = $conn->prepare($sql);
